drivers: fsi: sbefifo: Add OCC driver
Add a driver to control communications with the OCC on the POWER9. The
driver communicates with OCC SRAM using the FSI SBEFIFO driver.

The format of the communication is a command followed by response. The
sequence must be done atomically. The driver performs these operations
asynchronously, such that a write operation starts the command, and a
read will gather the response data.

Signed-off-by: Edward A. James <eajames@us.ibm.com>
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>