ARM: dts: aspeed: Update i2c binding
The I2C binding changed when upstreaming the I2C driver.

This adds aliases for the 14 built-in I2C busses to ensure userspace
sees the numbering that we had with the old driver.

This change moves the description of the i2c buses to the bottom of the
dtsi to improve readability.

OpenBMC-Staging-Count: 1
Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>