ARM: aspeed: Request NCSI mux
Garrison, Barreleye and Palmetto use the MAC in NCSI mode but were not
requesting the pins, but it happened to work. This change makes them
explicitly request the configuration they are using.

For Garrison, this also drops the muxing of ROMCS[1-3]# and ROMA24. I
assume they are unused as no other system designs have used them.

OpenBMC-Staging-Count: 1
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Reviewed-by: Xo Wang <xow@google.com>
[joel: Squished similar patches and added to commit message]
Signed-off-by: Joel Stanley <joel@jms.id.au>