arm: dts: aspeed: Update old style flash nodes
Reorganize flash controllers into the ast2400 device tree to match what
is used upstream. This moves the nodes from the machine dts to the
common dtsi.

This also affected the newly added Mellanox MSN device tree, which is
also updated in this series.

Barreleye wasn't booting with the new aspeed-smc driver.

Signed-off-by: Edward A. James <eajames@us.ibm.com>
[joel:
  - rework commit message
  - add missing m25p,fast-read properties
  - add pinmux for pnor
  - use upstream node names
     * bmc_pnor becomes fmc
     * host_pnor becomes spi
  - fix the facebook machine
]
Signed-off-by: Joel Stanley <joel@jms.id.au>
Reviewed-by: Cédric Le Goater <clg@kaod.org>
Reviewed-by: Eddie James <eajames@linux.vnet.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>