drivers/misc: add Aspeed LPC snoop driver
This driver enables the LPC snoop hardware on the ASPEED BMC
which generates an interrupt upon every write to an I/O port
by the host.

This is typically used to monitor BIOS boot progress by listening
to well-known debug port 80h.

The functionality in this commit just saves all snooped values
to a circular 2K buffer in the kernel, subsequent commits can
act on the values to do things with them.

OpenBMC-Staging-Count: 1
Signed-off-by: Robert Lippert <rlippert@google.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>