ARM: aspeed: fix reset of AHB controller
We have a reset of the AHB controller in the early boot path otherwise
some peripherals (i2c) are not accessible.

The reset register is level triggered (for all lines except the JTAG
peripheral): write 1 to reset, write 0 has no effect.

In c93456b24c68 ("aspeed: Don't blast SCU04 at boot time") we tried to
preserve the state of the JTAG controller, however, we inadvertently
reset a bunch of devices by doing the read/modify/write to the reset
register.

Preserve only the state of the JTAG reset.

Fixes: c93456b24c68 ("aspeed: Don't blast SCU04 at boot time")
Signed-off-by: Joel Stanley <joel@jms.id.au>