ARM: dts: aspeed: barreleye: Correct bad LED nodes
There was a bug in the definition of the heartbeat LED that meant we
were selecting the wrong GPIO for the driver. This fails now that we
have the correct pinmux settings for the Ethernet MAC.

When looking at the schematic for the correct number I found that the
other LEDs were mislabeled, and in the case of the power LED not
present at all.

Finally, the keep property is reinstated. This was present in an older
version of the dts as it is the required behaviour.

The LEDs are now accurate according to the Ingrasys document
0101DAW00-000-G, dated Wednesday, April 20, 2016, rev A00. The file
name is BARRELEYE-IO_BOARD-MP-HW-SCH-A00_20160419_FINAL.pdf

OpenBMC-Staging-Count: 1
Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>