dt-bindings: max31785: Add use-stored-presence property
It's a system design decision with regards to managing the default
values of registers such as FAN_CONFIG_1_2. Do not assume the hardware
is configured such as the fans are marked as installed, rather expose a
devicetree property to inform the driver of whether it should treat the
devicetree or the device state as canonical.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>