ARM: dts: aspeed: witherspoon: Add gpio-keys-polled for fan presence
Add gpio-keys-polled node for pca955x fan presence lines. Change the binding
of the pca9552 to remove the leds and instead define all the provided gpio pins
there. Add back the leds in a 'gpio-leds' compatible entry.

Polling period of 1 second to be determined to be the max acceptable
value based on discussions with Brad Bishop and Matt Spinler. Tested
with gpiomon program - verified.

Patch based on Cedric's example - thanks! Tested on witherspoon.

OpenBMC-Staging-Count: 1
Signed-off-by: Edward A. James <eajames@us.ibm.com>
Reviewed-by: Christopher Bostic <cbostic@linux.vnet.ibm.com>
Acked-by: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Joel Stanley <joel@jms.id.au>