ARM: dts: aspeed: Add ranges to SCU node
The SCU contains a number of devices. To date we have the pinctrl node,
which doesn't have a reg property, the out of tree clock drivers which
do have a single address in their reg properties, and the timeriomem
node which has an address and length.

From now on we will use a ranges property, allowing all of the devices
under the SCU to specify an address and size starting from the SCU base.

This allows the timeriomem node to live under the SCU syscon node. This change
is required to get the timeriomem driver to probe on the ast2400.

Reviewed-by: Rick Altherr <raltherr@google.com>
Acked-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>