ARM: dts: aspeed-romulus: Add I2C devices
Enable the buses that are in use and the devices that are attached.
Currently that is just the battery backed RTC.

Some of these buses are for hotplugged cards, such as PCIe cards. Others
do not yet have upstream drivers, so there are no devices attached.

Reviewed-by: Brendan Higgins <brendanhiggins@google.com>
Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>
(cherry picked from commit 11520916c817f88dd67b4e2cbd0d7409d2d529ee)
Signed-off-by: Joel Stanley <joel@jms.id.au>