ARM: dts: aspeed: Ensure platform mux RMII1 for MAC0
I noticed that Witherspoon was missing its pinctrl properties for the mac0
node. On auditing the rest of the DTS files, the MSN and Q71L platforms
appeared to have the same issue.

Add the pinctrl nodes to ensure MAC0 has RMII1 muxed. This provides mutual
exclusion with the GPIO subsystem ensuring broken userspace can't interfere
with networking.

OpenBMC-Staging-Count: 1
Cc: Mykola Kostenok <c_mykolak@mellanox.com>
Cc: Patrick Venture <venture@google.com>
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Acked-by: Mykola Kostenok <c_mykolak@mellanox.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>