fsi: sbefifo: Avoid using a timer to drive FIFO transfers
The timer subsystem appears to misbehave when requesting an immediate
expiry via `mod_timer(..., jiffies)`. The SBEFIFO driver uses this
pattern for scheduling work when it either adds a new transfer to an
empty queue or the timer callback finds the transfer queue is not empty.
In both cases work should be scheduled immediately, however the current
approach is not optimal:

The comment [0] outlines an issue where OCC sensor polling on the BMC
can come at remarkable cost. Delays of up to and above 10 seconds can be
experienced when reading an OCC attribute.

Tracing of the subsystems involved (FSI, the SBEFIFO, the OCC, the
scheduler and the timer subsystems) showed that the SBEFIFO timer
callbacks had a saw-tooth pattern of execution latency in jiffies
relative to their scheduled expiry time. In the measurements the typical
latency of an immediate reschedule was 15 jiffies, though this could be
as few as 1 or greater than 100 in the pathological case. Surprisingly,
these delays coincided in length with system idle time between the
SBEFIFO transfers - it appears timer subsystem is not properly
accounting for the immediacy of the request[1].

Querying sensors on the OCC incurs multiple SBEFIFO transfers, each with
its own (increasing) expiry latency, which leads to the poor
responsiveness observed above.

Stop misusing the timer infrastructure for work queue management and
incurring associated penalties (even if they appear to be a bug in the
timer core).  Instead, use a workqueue and use queue_work() and
queue_delayed_work() as required. With this change, reading an OCC
sensor typically takes ~0.5 seconds wall-clock in the un-cached case and
rarely exceeds 1 second. In the cached case the read completes in around
0.01 seconds.

Finally, using a workqueue enables the use of mutexes in place of
spinlocks, making way for future changes down the stack in the FSI core.

[0] https://github.com/openbmc/openbmc/issues/2696#issuecomment-351207046
[1] It seems scheduling a timer with `mod_timer(..., jiffies)` is a
    pathological case: Changing all such call-sites in the SBEFIFO driver to
    use `mod_timer(..., jiffies + 1)` yielded a similar improvement in
    performance to the workqueue strategy.

OpenBMC-Staging-Count: 1
Signed-off-by: Edward A. James <eajames@us.ibm.com>
[arj: Rework the commit message]
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>