clk: aspeed: Prevent reset if clock is enabled
According to the ASPEED specification, the reset and enable sequence
should be done when the clock is stopped. The specification doesn't
define behavior if the reset is done while the clock is enabled.

From testing on the AST2500 and AST2400, the LPC Controller has problems
if the clock is reset while enabled.

Therefore, check whether the clock is enabled or not before performing
the reset and enable sequence in the ASPEED clock driver.

OpenBMC-Staging-Count: 1
Signed-off-by: Eddie James <eajames@linux.vnet.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>